+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Upcoming
+* Fixed digger zombies overriding off hand items
+
+## 1.10.1
+* Launching creepers inaccuracy can now be configured
+* Skeleton Spammers now spam less, deal less damage and can be spawned with "enhancedai:spammer" tag
+
+## 1.10.0
+* Updated to 1.19.4
+
 ## 1.9.2
 * Walking fuse creepers now slowdown when exploding
 * Reduced animals base attack damage (4 -> 3)